[CONTINUUM-2666] added a check to prevent NPE
git-svn-id: https://svn.apache.org/repos/asf/continuum/trunk@1164024 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-xmlrpc/continuum-xmlrpc-server/src/main/java/org/apache/maven/continuum/xmlrpc/server/ContinuumServiceImpl.java
@@ -2221,7 +2221,10 @@ public class ContinuumServiceImpl
         ConfigurationService configurationService = continuum.getConfiguration();
         org.apache.continuum.configuration.BuildAgentGroupConfiguration buildAgentGroupConfiguration =
             configurationService.getBuildAgentGroup( name );
-        configurationService.removeBuildAgentGroup( buildAgentGroupConfiguration );
+        if ( buildAgentGroupConfiguration != null )
+        {
+            configurationService.removeBuildAgentGroup( buildAgentGroupConfiguration );
+        }
 
         return 0;
     }
